Add order query methods
