+++ src/index.ts
@@ -1,8 +1,9 @@
 import Web3 from "web3";
 import Web3Core from "web3-core";
 import { BasicMintMetadata, MintData, MintMetadata } from "./models/mint";
-import { Configuration, Order } from "./models/commons";
+import { Configuration } from "./models/commons";
 import { MatchEvent } from "./models/events";
+import { Order, OrderFilter } from "./models/orders";
 
 /**
  * Rarible SDK - Interface
@@ -60,4 +61,31 @@ export declare class RaribleSDK {
     data: MintData,
     metadata: BasicMintMetadata
   ): Promise<MintMetadata>;
+
+  /**
+   * Gets an Order given an order's hash.
+   *
+   * @param {string} hash - Hash of the order.
+   */
+  public getOrder(
+    hash: string
+  ): Promise<Order>;
+
+   /**
+   * Gets a Sell Order given a filter.
+   *
+   * @param {OrderFilter} filter - Defines criteria to filter orders by.
+   */
+  public getSellOrder(
+    filter: OrderFilter
+  ): Promise<Order>;
+
+  /**
+   * Gets a Buy Order given a filter.
+   *
+   * @param {OrderFilter} filter - Defines criteria to filter orders by.
+   */
+   public getBuyOrder(
+    filter: OrderFilter
+  ): Promise<Order>;
 }
